Version 1.01.4
ICobolCopybookTextSource interface and is implemented by ReadCobolCopybook class. This Interface can be used as a Cobol-Copybook Source

    Added method asCobolCopybookWalker to Cb2xml3 class. This create a CobolTree walker which calls a supplied listner with various Cobol Items
    Added/Updated classes to updated the Java classes (IITem, ICondition) representing Cobol-Copybook objects.

Interface IReadCobolCopybook has been replaced by ICobolCopybookTextSource. IReadCobolCopybook was a new interface in vesion 1.01.3 so this change should not effect many people.
Added a new Walker that walks through a Cobol copybook and calls a user supplied Listner.
Added updateCopybook method to class Cb2xml. This method sllows you to update the Cobol Item class in a Copybook class.